# docs/src/assets/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Message Vault HTTP API",
    "description": "HTTP API for a local Message Vault. Bearer session tokens come from login. API tokens come from Settings → Account.",
    "license": {
      "name": "Fair Core License 1.0 (ALv2 future)",
      "url": "https://github.com/bitrealm-io/message-vault/blob/main/LICENSE.md"
    },
    "version": "0.8.3"
  },
  "paths": {
...
    "/v1/saved-searches": {
      "get": {
        "tags": [
          "Saved searches"
        ],
        "summary": "List the account's saved searches, A–Z.",
        "operationId": "saved_searches_list_handler",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearchesListResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Saved searches"
        ],
        "summary": "Create a saved search: `201 Created`, `Location: /v1/saved-searches/{id}`,\nand the row.",
        "operationId": "saved_searches_create_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSearchBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Path of the new saved search"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearch"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/saved-searches/{id}": {
      "delete": {
        "tags": [
          "Saved searches"
        ],
        "summary": "Delete a saved search.",
        "description": "Deleting an import-created saved search removes the shortcut only. The\n`vault_imports` row it pointed at is the account's permanent record of that\nrun and is never touched here.",
        "operationId": "saved_searches_delete_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Saved search id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Saved search deleted"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Saved searches"
        ],
        "summary": "Replace a saved search's name and query, and return it.",
        "operationId": "saved_searches_update_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Saved search id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSearchBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearch"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Problem": {
        "type": "object",
        "description": "An RFC 7807 problem document: the body of every failure the vault answers,\nserved as `application/problem+json` (ADR-0010).\n\n`type` is the URL of the page describing this kind of failure, one page per\ntype, or `about:blank` for an internal error. A validation failure carries\n`errors`, every field that failed, in place of `detail`. Every problem\nrepeats the response's `x-request-id` as `request_id`, so the person\nreading the failure and the operator reading the log are looking at the\nsame request.\n\nThe extension members belong to one type each: `word` and `did_you_mean`\nto `search-query-invalid`, `retry_after` to `rate-limited`.",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "One sentence about this occurrence. Absent on a validation failure,\nwhich lists `errors` instead."
          },
          "did_you_mean": {
            "type": [
              "string",
              "null"
            ],
            "description": "`search-query-invalid`: a word the language does have, within a small\nedit distance of the one used."
          },
          "errors": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Every rule the request broke, one sentence each. Only on\n`validation-failed`."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The `x-request-id` of the response this came in."
          },
          "retry_after": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "`rate-limited`: seconds until an attempt may succeed, the same figure\nas the `Retry-After` header.",
            "minimum": 0
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The HTTP status, repeated in the body.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "The type's fixed, human-readable name."
          },
          "type": {
            "type": "string",
            "description": "URL of the page describing this kind of failure; `about:blank` for a\n`500 Internal Server Error`."
          },
          "word": {
            "type": [
              "string",
              "null"
            ],
            "description": "`search-query-invalid`: the `word:` the query used."
          }
        }
      },
...
      "SavedSearch": {
        "type": "object",
        "description": "One row of `saved_searches`.",
        "required": [
          "id",
          "name",
          "query",
          "kind"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Saved search id, unique across the vault."
          },
          "kind": {
            "type": "string",
            "description": "`manual` or `import`."
          },
          "name": {
            "type": "string",
            "description": "Display name, unique per account."
          },
          "query": {
            "type": "string",
            "description": "Query string, run against the conversation list."
          }
        }
      },
      "SavedSearchBody": {
        "type": "object",
        "description": "A saved search's name and query.",
        "required": [
          "name",
          "query"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "query": {
            "type": "string"
          }
        }
      },
      "SavedSearchesListResponse": {
        "type": "object",
        "description": "The account's saved searches, A–Z.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SavedSearch"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}